+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] (beta)
 
+### Fixed
+- Retry transient 400 "invalidRequest" errors during onedrive & sharepoint backup.
+
 ### Changed
 - When running `backup details` on an empty backup returns a more helpful error message.